User APIs
Small change to Job, Company and JobSave APIs

+++ src/main/java/com/jobplatform/repositories/UserRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<UserAccount,Integer> {
+public interface UserRepository extends JpaRepository<UserAccount,Long> {
     Optional<UserAccount> findByEmail(String email);
     Optional<UserAccount> findByResetPasswordToken(String token);
 }
